start split slider
