+++ .gitignore
@@ -6,4 +6,5 @@ node_modules/
 /img/.DS_Store
 /js/.DS_Store
 /js/slider/.DS_Store
+js/main-page/.DS_Store
 /style/.DS_Store
