+++ js/script.js
@@ -1,5 +1,5 @@
 require('./main-page/validator.js');
 require('./main-page/mask.js');
 require('./main-page/mobile-menu.js');
-require('./slider/slider-images.js');
-require('./slider/slider-pagination.js');
+require('./slider/show-images.js');
+require('./slider/next-prev.js');
